# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/pressings/tarif/": {
            "get": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Get all tarifs",
                "description": "Retrieve a list of all tarifs",
                "operationId": "getAllTarifs",
                "responses": {
                    "200": {
                        "description": "List of all tarifs"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Create a new tarif",
                "description": "Create a new tarif with the given details",
                "operationId": "createTarif",
                "requestBody": {
                    "description": "Tarif object to be created",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "New tarif created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "price": {
                                            "type": "number"
                                        },
                                        "id_service": {
                                            "type": "integer"
                                        },
                                        "id_article": {
                                            "type": "integer"
                                        },
                                        "id_pressing": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/pressings/tarif/{id}": {
            "get": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Get a Tarif by ID",
                "description": "Returns a Tarif object corresponding to the provided ID",
                "operationId": "getTarifById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tarif to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Tarif not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Update a Tarif by ID",
                "description": "Updates a Tarif object corresponding to the provided ID",
                "operationId": "updateTarifById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tarif to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "The new information for the Tarif",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Tarif not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pressing"
                ],
                "summary": "Delete a Tarif by ID",
                "description": "Deletes a Tarif object corresponding to the provided ID",
                "operationId": "deleteTarifById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Tarif to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Tarif not found"
                    }
                }
            }
        }
    }
}